Abilities largely complete but also untested.
NOTE: NO CHECKS TO SEE IF A SHIP CAN USE AN ABILITY ARE MADE YET!
The ship class needs to be able to report whether certain actions are currently possible and the abilities class needs to use such new ship methods.  This will likely mean that a shipID be passed as an argument to some ability functions.